requirements.txt
fastapi
openai
uvicorn
python-dotenv
langchain
langchain_community
langchain_experimental
langchain-core
langchain-openai
tiktoken
pinecone
langchain-pinecone 
pinecone-notebooks
unstructured 
openpyxl
docx2txt
python-multipart
pypdf